SW1/3.RAG-LC2/requirements.txt
langchain-core
langchain-text-splitters
langchain-mistralai
langchain-huggingface
langchain-community
langgraph
faiss-cpu
python-dotenv
pypdf